requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
fastapi==0.71.0
frozenlist==1.2.0
future==0.18.2
idna==3.3
line-bot-sdk==2.0.1
multidict==5.2.0
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
